requirements.txt
fastapi 
uuid 
uvicorn 
gunicorn 
WebSockets 
python-dotenv 
aioredis